refactor: create a component for list tasks
