+++ web/src/pages/index.tsx
@@ -1,18 +1,6 @@
 import { Filters } from "@/components/Filters/filters";
 import { Form } from "@/components/Form/form";
-import { Task } from "@/components/task";
-
-const mockTasks = [
-  { id: "1", title: "Comprar pão", done: false, createdAt: "24/04/2024" },
-  { id: "2", title: "Lavar o carro", done: true, createdAt: "24/04/2024" },
-  {
-    id: "3",
-    title: "Estudar para o exame",
-    done: false,
-    createdAt: "24/04/2024",
-  },
-  { id: "4", title: "Fazer exercícios", done: false, createdAt: "24/04/2024" },
-];
+import { TaskList } from "@/components/TaskList/task-list";
 
 export default function Home() {
   return (
@@ -20,17 +8,7 @@ export default function Home() {
       <h1 className="text-2xl font-bold mb-4">Minhas Tarefas</h1>
       <Form />
       <Filters />
-
-      <div className="space-y-2">
-        {mockTasks.map((task) => (
-          <Task
-            key={task.id}
-            id={task.id}
-            title={task.title}
-            createdAt={task.createdAt}
-          />
-        ))}
-      </div>
+      <TaskList />
     </main>
   );
 }
